# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Vue3 Laravel Admin",
        "version": "0.0.1",
        "description": "管理 API。成功包络 code=0000；失败使用对应 HTTP 状态及业务 code。写入管理数据须提交最新列表 version，冲突返回 409。Web 登录使用 HttpOnly 刷新 Cookie，刷新及退出须带 X-CSRF-Token；desktop 使用响应中的 refreshToken。所有统计按实时权限过滤，审计时间使用 UTC。"
    },
    "servers": [{ "url": "/api" }],
    "security": [{ "bearerAuth": [] }],
    "paths": {
...
        "/v1/auth/refresh": {
            "post": {
                "operationId": "auth.refresh",
                "tags": ["Auth"],
                "parameters": [
                    {
                        "name": "X-CSRF-Token",
                        "in": "header",
                        "description": "Web 客户端必填，使用登录响应中的 csrfToken；desktop 无需此字段。",
                        "schema": { "type": "string" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object",
                                        "properties": {
                                            "client": { "type": "string", "const": "web" },
                                            "refreshToken": { "type": ["string", "null"], "maxLength": 100 }
                                        },
                                        "required": ["client"]
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "client": { "type": "string", "const": "desktop" },
                                            "refreshToken": { "type": "string", "minLength": 1, "maxLength": 100 }
                                        },
                                        "required": ["client", "refreshToken"]
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": { "type": "string", "const": "0000" },
                                        "msg": { "type": "string" },
                                        "data": { "$ref": "#/components/schemas/Tokens" }
                                    },
                                    "required": ["code", "msg", "data"]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未登录、令牌过期或会话撤销",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
                    },
                    "403": {
                        "description": "无权限或来源/CSRF 校验失败",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
                    },
                    "404": {
                        "description": "记录不存在",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
                    },
                    "409": {
                        "description": "版本冲突或受保护的数据",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
                    },
                    "422": {
                        "description": "输入校验失败",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
                    },
                    "429": {
                        "description": "请求过于频繁",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
                    },
                    "500": {
                        "description": "服务暂时不可用",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
                    }
                },
                "security": [{}]
            }
        },
...
        "/v1/auth/profile": {
            "put": {
                "operationId": "auth.profile",
                "tags": ["Auth"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": { "type": "string", "maxLength": 100 },
                                    "currentPassword": { "type": "string" },
                                    "password": { "type": ["string", "null"], "minLength": 12, "maxLength": 128 },
                                    "password_confirmation": {
                                        "type": ["string", "null"],
                                        "minLength": 12,
                                        "maxLength": 128
                                    }
                                },
                                "required": ["name"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": { "type": "string", "const": "0000" },
                                        "msg": { "type": "string" },
                                        "data": { "$ref": "#/components/schemas/Identity" }
                                    },
                                    "required": ["code", "msg", "data"]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未登录、令牌过期或会话撤销",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
                    },
                    "403": {
                        "description": "无权限或来源/CSRF 校验失败",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
                    },
                    "404": {
                        "description": "记录不存在",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
                    },
                    "409": {
                        "description": "版本冲突或受保护的数据",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
                    },
                    "422": {
                        "description": "输入校验失败",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
                    },
                    "429": {
                        "description": "请求过于频繁",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
                    },
                    "500": {
                        "description": "服务暂时不可用",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } }
                    }
                },
                "security": [{ "bearerAuth": [] }]
            }
        },
...
    },
    "components": {
        "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } },
        "schemas": {
...
            "ApiError": {
                "type": "object",
                "properties": {
                    "code": { "type": "string" },
                    "msg": { "type": "string" },
                    "data": { "type": "null" },
                    "errors": {
                        "type": ["object", "null"],
                        "additionalProperties": { "type": "array", "items": { "type": "string" } }
                    }
                },
                "required": ["code", "msg", "data"],
                "title": "ApiError"
            },
...
            "Identity": {
                "type": "object",
                "properties": {
                    "userId": { "type": "string" },
                    "userName": { "type": "string" },
                    "email": { "type": "string", "format": "email" },
                    "roles": { "type": "array", "items": { "type": "string" } },
                    "buttons": { "type": "array", "items": { "type": "string" } },
                    "accessVersion": { "type": "integer" }
                },
                "required": ["userId", "userName", "email", "roles", "buttons", "accessVersion"],
                "title": "Identity"
            },
...
            "Tokens": {
                "type": "object",
                "properties": {
                    "token": { "type": "string" },
                    "refreshToken": {
                        "type": "string",
                        "description": "仅 desktop 客户端返回；Web 使用 HttpOnly Cookie"
                    },
                    "csrfToken": { "type": "string" },
                    "expiresIn": { "type": "integer", "description": "访问令牌有效期，单位秒" }
                },
                "required": ["token", "csrfToken", "expiresIn"],
                "title": "Tokens"
            },
...
        },
...
    }
}